% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rats.R
\name{alloc_out}
\alias{alloc_out}
\title{Create output structure.}
\usage{
alloc_out(annot)
}
\arguments{
\item{annot}{a dataframe with at least 2 columns: target_id and parent_id.}
}
\value{
a list-based structure of class dtu.
}
\description{
Create output structure.
}

