% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot2ids.R
\name{annot2ids}
\alias{annot2ids}
\title{Extract matching transcript and gene IDs from a GTF file.}
\usage{
annot2ids(annotfile, transc_header = "target_id", gene_header = "parent_id")
}
\arguments{
\item{annotfile}{A GTF file.}

\item{transc_header}{The tile for the transcripts column in the output. (target_id)}

\item{gene_header}{The tile for the genes column in the output. (parent_id)}
}
\value{
A data.frame with two columns, matching transcript IDs to gene IDs.
}
\description{
This function performs no file format validation. Please use the appropriate type.
It is NOT compatible GFF3, as there is no rigid specification for ID attributes.
}

