% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{annot2models}
\alias{annot2models}
\title{Prepare GTF for gene model plotting.}
\usage{
annot2models(annotfile, threads = 1L)
}
\arguments{
\item{annotfile}{A GTF file.}

\item{threads}{Number of processing threads (Default 1).}
}
\value{
A list (by gene) of GRangeLists (by transcript).
}
\description{
Create GRanges objects for each transcript and bind them into a GRangeList for each gene.
Then one can easily \code{ggbio::autoplot()} a gene's models by its gene ID.
}
