% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{annotation_inconsistent}
\alias{annotation_inconsistent}
\title{Feature IDs match between}
\usage{
annotation_inconsistent(annot, tids)
}
\arguments{
\item{annot}{Annotation dataframe.}

\item{tids}{List of transcript ID vectors.}
}
\value{
bool
}
\description{
This is applied after bootstrapped counts are checked and combined down to a single vector of values per sample.
So it covers both bootstrapped and non-bootstrapped input.
}
