% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{At_250genes_50singletx_200multitx_513totaltx}
\alias{At_250genes_50singletx_200multitx_513totaltx}
\title{Small example sleuth object containing transcript abundance estimates and bootstrap data}
\format{A sleuth object.}
\usage{
At_250genes_50singletx_200multitx_513totaltx
}
\description{
\href{http://combine-lab.github.io/salmon/}{Salmon} abundance estimates and bootstraps were
calculated for 513 Arabidopsis thaliana transcripts from the TAIR10 annotation. These
transcripts map to 250 genes - 50 ofwhich are single-transcript genes, 200 of which have
multiple transcripts (463 in total).
}
\details{
The data were computed for 3 biological replicates in 2 conditions:
\itemize{
  \item{Col: 3 replicates}
  \item{Vir: 3 replicates}
}

The \href{http://combine-lab.github.io/salmon/}{salmon} abundance information is then
processed with the routine \code{\link{gen_example_sleuth}} (including a convertion to
the correct hd5 format if required) to generate theoutput sleuth object that is saved here.

See the help page for \code{\link{At_RNAMeth_sleuth}} for a description of the sleuth
object structure.
}
\keyword{datasets}

