% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{At_tair10_t2g}
\alias{At_tair10_t2g}
\title{{Arabidopsis thaliana} TAIR10 annotation information}
\format{A data frame with 41,671 observations of 3 variables
\describe{
  \item{target_id}{ENSEMBL transcript id.}
  \item{parent_id}{ENSEMBL gene id.}
  \item{other_id_1}{ENSEMBL external id.}

}}
\usage{
At_tair10_t2g
}
\description{
Annotation information extracted from ensembl with \code{\link{biomaRt}} that links
transcript level annotations with the parent gene level annotations and an external
annotation.
}
\details{
The data were generated with: \preformatted{
library("biomaRt")
mart <- biomaRt::useMart(biomart = "plants_mart", host="plants.ensembl.org", dataset = "athaliana_eg_gene")
At_tair10_t2g <- biomaRt::getBM(attributes = c("ensembl_transcript_id", "ensembl_gene_id", "external_gene_id"), mart = mart)
At_tair10_t2g <- dplyr::rename(t2g, target_id = ensembl_transcript_id, parent_id = ensembl_gene_id, other_id_1 = external_gene_id)
devtools::use_data(At_tair10_t2g)
}
}
\keyword{datasets}

