% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\docType{class}
\name{BarProgressUpdate-class}
\alias{BarProgressUpdate}
\alias{BarProgressUpdate-class}
\title{An S4 progress bar update class, subclassed from ProgressUpdate}
\description{
An S4 progress bar update class, subclassed from ProgressUpdate
}
\section{Slots}{

\describe{
\item{\code{pb}}{The current progress bar}
}}

