% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rats.R
\name{calculate_stats}
\alias{calculate_stats}
\title{Calculate statistics across the columns of a dataframe}
\usage{
calculate_stats(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
A dataframe containing the calculated statistics, one in each column
}
\description{
Calculate statistics across the columns of a dataframe
}

