% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rats.R
\name{calculate_stats2}
\alias{calculate_stats2}
\title{Calculate row-wise statistics in a dataframe.}
\usage{
calculate_stats2(df)
}
\arguments{
\item{df}{A dataframe.}
}
\value{
A dataframe containing the sum, mean, variance per row.

It expects and preserves rownames.
}
\description{
Calculate row-wise statistics in a dataframe.
}

