% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rats.R
\name{calculate_tx_proportions}
\alias{calculate_tx_proportions}
\title{Calculate the proportion of counts which are assigned to each transcript in a gene}
\usage{
calculate_tx_proportions(sleuth_data, transcripts, counts_col = "est_counts")
}
\arguments{
\item{sleuth_data}{a sleuth object}

\item{transcripts}{a dataframe listing the transcripts to process, and their parent genes,
with at least column \code{target_id} & \code{parent_id}}

\item{counts_col}{the sleuth column to use for the calculation (est_counts or tpm), default est_counts}
}
\description{
Calculate the proportion of counts which are assigned to each transcript in a gene
}

