% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rats.R
\name{denest_boots}
\alias{denest_boots}
\title{Extract bootstrap counts into a less nested structure.}
\usage{
denest_boots(wcl, slo, tx, samples, COUNTS_COL, BS_TARGET_COL)
}
\arguments{
\item{wcl}{A cluster of workers (threads).}

\item{slo}{A sleuth object.}

\item{tx}{A vector of transcript ids.}

\item{samples}{A numeric vector of samples to extract counts for.}

\item{COUNTS_COL}{The name of the column with the counts.}

\item{BS_TARGET_COL}{The name of the column with the transcript IDs.}
}
\value{
a list of data.tables, one per sample, containing all the bootstrap counts.
}
\description{
NA replaced with 0.
}

