% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{do_boot}
\alias{do_boot}
\title{Perform the bootstraps.}
\usage{
do_boot(lean, tx_filter, eltr, elge, boot_data_A = NULL,
  boot_data_B = NULL, count_data_A = NULL, count_data_B = NULL,
  niter = NULL, threads = 1, verbose = TRUE, test_transc = TRUE,
  test_genes = TRUE, ...)
}
\arguments{
\item{lean}{Minimal insight reported.}

\item{tx_filter}{The annotation.}

\item{eltr}{Selection vector for eligible transcripts.}

\item{elge}{Selection vector for eligible genes.}

\item{boot_data_A}{Bootstrapped quantifications for first condition.}

\item{boot_data_B}{Bootstrapped quantifications for second condition.}

\item{count_data_A}{Plain quantifications for first condition.}

\item{count_data_B}{Plain quantifications for second condition.}

\item{niter}{Number of iterations (for quantification bootstraps only).}

\item{threads}{Number of threads.}

\item{verbose}{Progress messages.}

\item{test_transc}{Do transcript-level test.}

\item{test_genes}{Do gene-level test.}

\item{...}{Other arguments to pass on to calculate_DTU().}
}
\description{
Abstracts the bootstrap process from the specific field names of each bootstrap
type and removes the associated code duplication.
}
\details{
Providing bootstrapped data will bootstrap the quantifications, whereas plain
data will bootstrap the replicates. Do not provide both.
}
