% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rats.R
\name{do_tests}
\alias{do_tests}
\title{Set up and execute the tests.}
\usage{
do_tests(counts_A, counts_B, tx_filter, testmode, full, count_thresh,
  correction)
}
\arguments{
\item{counts_A}{A data.table of counts for condition A. x: sample, y: transcript.}

\item{counts_B}{A data.table of counts for condition B. x: sample, y: transcript.}

\item{tx_filter}{A data.table with target_id and parent_id.}

\item{testmode}{One of "G-test", "prop-test", "both".}

\item{full}{Either "full" (for complete output structure) or "short" (for bootstrapping).}

\item{count_thresh}{Minimum number of counts per replicate.}

\item{correction}{Multiple testing correction type.}
}
\value{
list
}
\description{
Set up and execute the tests.
}

