% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{dtu_plurality_summary}
\alias{dtu_plurality_summary}
\title{Summary of DTU plurality.}
\usage{
dtu_plurality_summary(dtuo)
}
\arguments{
\item{dtuo}{A DTU object.}
}
\value{
A data.frame with category names in the first column and values in the second column.
}
\description{
A tally of genes based on how many isoforms show significant change.
}
