% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{dtu_summary}
\alias{dtu_summary}
\title{Summary of DTU calling.}
\usage{
dtu_summary(dtuo)
}
\arguments{
\item{dtuo}{a DTU object.}
}
\value{
named numerical vector giving a quick overview of the results
}
\description{
Summary of DTU calling.
}

