% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{dtu_switch_summary}
\alias{dtu_switch_summary}
\title{Summary of isoform switching events.}
\usage{
dtu_switch_summary(dtuo)
}
\arguments{
\item{dtuo}{A DTU object.}
}
\value{
A data.frame with catefory names in the first column and values in the second column.
}
\description{
A tally of genes showing isoform switching.
}
