% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{dtuSummary}
\alias{dtuSummary}
\title{Summary of DTU}
\usage{
dtuSummary(dtuo)
}
\arguments{
\item{dtuo}{a DTU object.}
}
\value{
named numerical vector giving a quick overview of the results
}
\description{
Summary of DTU
}

