% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rats.R
\name{filter_and_match2}
\alias{filter_and_match2}
\title{Extract filtered counts, assuming the bootstraps and replicates are identical in their
order, number and presence of transcripts.}
\usage{
filter_and_match2(kal, tx_filter = c(TRUE), counttype = "tpm")
}
\arguments{
\item{kal}{A list of kallisto objects.}

\item{tx_filter}{A logic vector to select bootstap rows (the same ones across all bootstraps).}

\item{counttype}{"tpm" or "est_counts".}
}
\description{
Extract filtered counts, assuming the bootstraps and replicates are identical in their
order, number and presence of transcripts.
}

