% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{g.test.1}
\alias{g.test.1}
\title{Log-likelihood test of goodness of fit.}
\usage{
g.test.1(obsx, px)
}
\arguments{
\item{obsx}{a numeric vector of finite positive counts, with at least one non-zero value.}

\item{px}{a vector of probabilities of the same length as xobs. ( sum(px) <= 1 )

The order of values in the two vectors should be the same.
If any value of xp is zero and the corresponding xobs is not, g.test.1 will always reject the hypothesis.
No corrections are applied.
No input checks are applied.}
}
\description{
For a set of observations against a set of probabilities.
}
\details{
General utility function.
}
