% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{g.test.n}
\alias{g.test.n}
\title{Log-likelihood test of independence.}
\usage{
g.test.n(repx, repy)
}
\arguments{
\item{repx}{a numeric matrix of positive counts, r x c (replicates by categories), with at least one non-zero value.}

\item{repy}{a numeric matrix of positive counts, r x c (replicates by categories), with at least one non-zero value.

The number and order of categories (columns) should be the same in both matrices.
No corrections are applied. 
No input checks are applied, as RATs needs to run this millions of times.}
}
\description{
For any number of sets of observations.
}
