% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{g.test}
\alias{g.test}
\title{Log-likelihood test of goodness of fit.}
\usage{
g.test(x, p = rep(1/length(x), length(x)))
}
\arguments{
\item{x}{a numeric vector of positive numbers, with at least one non-zero value.}

\item{p}{a vector of probabilities of the same length of x.

Sourced and adapted from from:
V3.3 Pete Hurd Sept 29 2001. phurd@ualberta.ca
http://www.psych.ualberta.ca/~phurd/cruft/g.test.r}
}
\description{
Log-likelihood test of goodness of fit.
}
