% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{gen_example_sleuth}
\alias{gen_example_sleuth}
\title{A function that generates the example sleuth object storred in
At_250genes_50singletx_200multitx_513totaltx.rda from the raw salmon 0.5.1. data}
\usage{
gen_example_sleuth(overwrite = FALSE, annot = "TAIR10",
  path = "/cluster/gjb_lab/nschurch/Projects/Arabidopsis_RNAMeth/salmon/rats_testdata_0.5.1_bootstraps")
}
\arguments{
\item{overwrite}{Logical flag for triggering an overwrite of At_250genes_50singletx_200multitx_513totaltx.rda.}

\item{annot}{name of the annotation used for the data - assumes this is part of the path to the files.}

\item{path}{root path to the raw salmon data - combined with annot to get the data for all replicates.}
}
\description{
A function that generates the example sleuth object storred in
At_250genes_50singletx_200multitx_513totaltx.rda from the raw salmon 0.5.1. data
}

