% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{get_plurality_ids}
\alias{get_plurality_ids}
\title{List the gene IDs by number of isoforms showing significant change.}
\usage{
get_plurality_ids(dtuo)
}
\arguments{
\item{dtuo}{A DTU object.}
}
\value{
A named numerical vector giving a tally of the results.
}
\description{
Get the IDs of DTU genes organised by the number of isoforms affected. This
is possible only if transcript-level results are enabled.
}
