% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{get_switch_ids}
\alias{get_switch_ids}
\title{List of genes that switch isoform ranks.}
\usage{
get_switch_ids(dtuo)
}
\arguments{
\item{dtuo}{A DTU object.}
}
\value{
A named list of character vectors.
}
\description{
Get the IDs of DTU genes where isoform rank switching occurs.
Switches of primary and non-primary isoforms are listed separately.
}
