% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{granges2ids}
\alias{granges2ids}
\title{Extract matching transcript and gene IDs from a GRanges object.}
\usage{
granges2ids(annot, transc_header = "target_id",
  gene_header = "parent_id")
}
\arguments{
\item{annot}{A GRanges object that was imported from a GTF file and retains all the metadata columns.}

\item{transc_header}{The title for the transcripts column in the output. (target_id)}

\item{gene_header}{The title for the genes column in the output. (parent_id)}
}
\value{
A data.table with two columns, matching transcript IDs to gene IDs.
}
\description{
The GRanges must have at least the following metadata columsn: `gene_id` and `transcript_id` (such as GRanges imported from GTF).
GFF3-style metadata columns are currently not supported.
}
