% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{gtf2ids}
\alias{gtf2ids}
\title{Extract matching transcript and gene IDs from a GTF file.}
\usage{
gtf2ids(annotfile, transc_header = "target_id",
  gene_header = "parent_id")
}
\arguments{
\item{annotfile}{A GTF file.}

\item{transc_header}{The title for the transcripts column in the output. (target_id)}

\item{gene_header}{The title for the genes column in the output. (parent_id)}
}
\value{
A data.table with two columns, matching transcript IDs to gene IDs.
}
\description{
This was previously called `annot2ids()`. The change to a more explicitly descriptive name follows the addition of
other annotation-related helper functions. Support for the old name will eventually be withdrawn.
}
\details{
GFF3 is currently not supported.
}
