% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{infer_bootnum}
\alias{infer_bootnum}
\title{Rule-of-thumb number of iterations.}
\usage{
infer_bootnum(slo, boot_data_A, boot_data_B)
}
\arguments{
\item{slo}{Sleuth object.}

\item{boot_data_A}{List of tables of bootstrapped counts.}

\item{boot_data_B}{List of tables of bootstrapped counts.}
}
\value{
The least number of iterations seen in the input.
}
\description{
Rule-of-thumb number of iterations.
}

