% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rats.R
\name{init_progress}
\alias{init_progress}
\title{Initialise progress updates}
\usage{
init_progress(on)
}
\arguments{
\item{on}{Flag indicating whether updates are on (TRUE) or not (FALSE)}
}
\value{
The progress update object
}
\description{
Initialise progress updates
}

