% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\docType{methods}
\name{initialize,BarProgressUpdate-method}
\alias{initialize,BarProgressUpdate-method}
\title{Initialisation of subclass}
\usage{
\S4method{initialize}{BarProgressUpdate}(.Object, ...)
}
\arguments{
\item{.Object}{base class}

\item{...}{unnamed arguments}
}
\description{
Initialisation of subclass
}

