% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\docType{methods}
\name{insert_steps,ProgressUpdate-method}
\alias{insert_steps,ProgressUpdate-method}
\title{Insert steps into progress update object}
\usage{
\S4method{insert_steps}{ProgressUpdate}(theObject, newsteps)
}
\arguments{
\item{theObject}{base class}

\item{newsteps}{dataframe of steps to insert into current steps dataframe}
}
\description{
Insert steps into progress update object
}

