% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rats.R
\name{mark_sibling_targets}
\alias{mark_sibling_targets}
\title{Compute a logical vector filter marking single-target parents in a data frame.}
\usage{
mark_sibling_targets(df)
}
\arguments{
\item{df}{a data frame with at least two variables, \code{target_id} & \code{parent_id}}
}
\description{
Compute a logical vector filter marking single-target parents in a data frame.
}

