% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rats.R
\name{mark_sibling_targets3}
\alias{mark_sibling_targets3}
\title{Compute a logical vector marking as FALSE the single-target parents in a data frame.}
\usage{
mark_sibling_targets3(ids, p2t)
}
\arguments{
\item{ids}{a data frame with at least two variables, \code{target_id} & \code{parent_id}.}

\item{p2t}{a list of vectors, listing the \code{target_id}s per \code{parent_id}.}
}
\value{
data.frame An updated version of the input ids.
}
\description{
Compute a logical vector marking as FALSE the single-target parents in a data frame.
}

