% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rats.R
\name{mark_singleparent_ids2}
\alias{mark_singleparent_ids2}
\title{Compute a logical vector filter marking multi-child parent IDs in a data frame.}
\usage{
mark_singleparent_ids2(ids)
}
\arguments{
\item{ids}{a data frame with at least two variables, \code{target_id} & \code{parent_id}

This is a second implementation of this routine. Faster, but less general.}
}
\description{
Compute a logical vector filter marking multi-child parent IDs in a data frame.
}

