% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{maxabs}
\alias{maxabs}
\title{Get largest value by absolute comparison.}
\usage{
maxabs(v)
}
\arguments{
\item{v}{A numeric vector.}
}
\value{
A numeric value.
}
\description{
General utility function.
}
\details{
Returns the original signed value with the largest absolute value in the vector. 
In case of equal absolutes between a positive and negative value, the positive 
value will always be the one that is returned.
}
