% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rats.R
\name{mean_and_var}
\alias{mean_and_var}
\title{Calculate mean and variance across the columns of a row}
\usage{
mean_and_var(row)
}
\arguments{
\item{row}{the row}
}
\value{
A vector containing the mean and variance
}
\description{
Calculate mean and variance across the columns of a row
}

