% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{parameters_are_good}
\alias{parameters_are_good}
\title{Check input parameters.}
\usage{
parameters_are_good(slo, annot, name_A, name_B, varname, COUNTS_COL, correction,
  p_thresh, TARGET_COL, PARENT_COL, BS_TARGET_COL, count_thresh, testmode,
  qboot, qbootnum, dprop_thresh, count_data_A, count_data_B, boot_data_A,
  boot_data_B, qrep_thresh, threads, rboot, rrep_thresh, rrep_as_crit)
}
\arguments{
\item{slo}{Sleuth object}

\item{annot}{Annotation dataframe.}

\item{name_A}{Condition name.}

\item{name_B}{Condition name.}

\item{varname}{Name of condition variable.}

\item{COUNTS_COL}{Name of counts column in bootstrap.}

\item{correction}{P-value correction method.}

\item{p_thresh}{Significance level.}

\item{TARGET_COL}{Name of transcript id column in annotation.}

\item{PARENT_COL}{Name of gene id column in annotation.}

\item{BS_TARGET_COL}{Name of transcript id column in bootstrap.}

\item{count_thresh}{Minimum transcript abundance per sample.}

\item{testmode}{Which test to run.}

\item{qboot}{Whether to bootstrap against quantifications.}

\item{qbootnum}{Number of bootstrap iterations.}

\item{dprop_thresh}{Minimum change in proportion.}

\item{count_data_A}{A dataframe of estimated counts.}

\item{count_data_B}{A dataframe of estimated counts.}

\item{boot_data_A}{A list of dataframes, one per sample, each with all the bootstrapped estimetes for the sample.}

\item{boot_data_B}{A list of dataframes, one per sample, each with all the bootstrapped estimetes for the sample.}

\item{qrep_thresh}{Confidence threshold.}

\item{threads}{Number of threads.}

\item{rboot}{Whether to bootstrap against samples.}

\item{rrep_thresh}{Confidence threshold.}

\item{rrep_as_crit}{Whether to use rrep as a DTU criterion.}
}
\value{
List: \itemize{
 \item{"error"}{logical}
 \item{"message"}{string}
 \item{"maxboots"}{Infered rule-of-thumbs number of iterations to do.}
 \item{"warn"}{logical}
 \item{"warnings}{list of strings}
}
}
\description{
Check input parameters.
}
