% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rats.R
\name{parameters_good}
\alias{parameters_good}
\title{Check input parameters.}
\usage{
parameters_good(sleuth_data, transcripts, ref_name, comp_name, varname,
  counts_col, correction, p_thresh, TARGET_ID, PARENT_ID, BS_TARGET_ID, verbose,
  threads)
}
\value{
List with a logical value and a message.
}
\description{
Check input parameters.
}

