% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rats.R
\name{parent_to_targets}
\alias{parent_to_targets}
\title{Group targets by parent.}
\usage{
parent_to_targets(ids)
}
\arguments{
\item{ids}{A dataframe with \code{target_id} and \code{parent_id}.}
}
\value{
A list of vectors: one vector for each parent, containing the repsective targets.
}
\description{
Group targets by parent.
}

