% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangedat-gen.R
\name{plot_range}
\alias{plot_range}
\title{Plot relationship of parameters.}
\usage{
plot_range(data)
}
\arguments{
\item{x, }{y parameter names to plot on x, y axes}

\item{grd}{parameter name to represent as grid}

\item{colour}{parameter name to represent as colour}
}
\description{
Plot relationship of parameters.
}

