% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot_shiny_volcano}
\alias{plot_shiny_volcano}
\title{Interactive volcano plot, using shiny.}
\usage{
plot_shiny_volcano(dtuo)
}
\arguments{
\item{dtuo}{A DTU object.}
}
\description{
This depends on the shiny package.
}
