% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangedat-gen.R
\name{plot_sim}
\alias{plot_sim}
\title{Plot relationship of parameters.}
\usage{
plot_sim(data, type = 1)
}
\arguments{
\item{data}{the output from combine_sim_dtu()}

\item{type}{1: dtu, 2: pval_AB, 3: pval_BA}
}
\description{
Plot relationship of parameters.
}

