% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangedat-gen.R
\name{plot_sim2}
\alias{plot_sim2}
\title{Plot relationship of parameters.}
\usage{
plot_sim2(data, type = "AvBvM")
}
\arguments{
\item{data}{the output from combine_sim_dtu()}

\item{type}{(str) "AvBvM", "A/BvM"}
}
\description{
Plot relationship of parameters.
}

