% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotDTU}
\alias{plotDTU}
\title{Plot DTU results.}
\usage{
plotDTU(dtuo, type = "propVmag")
}
\arguments{
\item{dtuo}{A DTU object.}

\item{type}{Type of plot: "propVcount", "dpropVcount", "propfoldVsig", "dpropVsig", "dprop"}
}
\value{
a ggplot2 object. Simply display it or you can also customize it.
}
\description{
Plot DTU results.
}

