% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotGeneDTU}
\alias{plotGeneDTU}
\title{Plot count or proportion changes for all transcripts of a specified gene.}
\usage{
plotGeneDTU(dtuo, pid, type = "counts", style = "bar")
}
\arguments{
\item{dtuo}{a DTU object.}

\item{pid}{a \code{parent_id} to make the plot for.}

\item{type}{a switch for plotting either "counts" or "proportion". Default "counts".}

\item{style}{style of plot. Either "bars" or "lines". Default "bars".}
}
\description{
Plot count or proportion changes for all transcripts of a specified gene.
}

