% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\docType{class}
\name{ProgressUpdate-class}
\alias{ProgressUpdate}
\alias{ProgressUpdate-class}
\title{An S4 base class to represent a progress update object.}
\description{
An S4 base class to represent a progress update object.
}
\section{Slots}{

\describe{
\item{\code{steps}}{A dataframe containing the progress positions (10%,20% etc) and associated text}

\item{\code{max}}{The maximum value of the progress positions}

\item{\code{on}}{Whether to run the progress update or not, default=FALSE}
}}

