% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangedat-gen.R
\name{rangedat_gen}
\alias{rangedat_gen}
\title{Generate artificial read-count data for simulated 2-transcript genes, 
systematically covering a broad range of proportions, magnitudes and fold-changes.}
\usage{
rangedat_gen(propfrom = 0, propto = 1, propby = 0.01, magnitudes = c(1,
  5, 10, 30, 100, 500, 1000), foldfrom = 0, foldto = 10, foldby = 0.5)
}
\arguments{
\item{propfrom, }{propto, propby: Range and step of proportions for 1st transcript. Sibling transcript complemented from these.}

\item{foldfrom, }{foldto, foldby: Range of transcript ratio fold changes.}

\item{magnitudes:}{vector of read-count sizes to use.}
}
\value{
list containing: \code{data} a minimal sleuth-like object.
                         \code{anno} a dataframe matching \code{target_id} to \code{parent_id}.
                         \code{sim} a list of dataframes with the simulation parameters per transcript per condition.
}
\description{
Generate artificial read-count data for simulated 2-transcript genes, 
systematically covering a broad range of proportions, magnitudes and fold-changes.
}

