% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{scale_data}
\alias{scale_data}
\title{Apply scaling.}
\usage{
scale_data(scaling, threads, steps, count_data_A, count_data_B,
  boot_data_A, boot_data_B)
}
\arguments{
\item{scaling}{Scaling factor(s).}

\item{threads}{Threads.}

\item{steps}{Infered input type.}

\item{count_data_A}{Non-bootstrapped data for one condition.}

\item{count_data_B}{Non-bootstrapped data for other condition.}

\item{boot_data_A}{Boostrapped data for one condition.}

\item{boot_data_B}{Bootstrapped data for other condition.}
}
\value{
List.
}
\description{
Apply scaling.
}
