% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulators.R
\name{sim_count_data}
\alias{sim_count_data}
\title{Generate an artificial dataset of bootstrapped abundance estimates, for code-testing or examples.}
\usage{
sim_count_data(cnames = c("A", "B"))
}
\arguments{
\item{cnames}{A vector of (two) name values for the comparison variable.}
}
\value{
A list with 3 elements. First, a data.frame with the corresponding annotation table. 
Then, 2 data.tables, one per condition. Each data table contains the abundance estimates for all the samples for the respective condition.
}
\description{
Based on sim_sleuth_data().
}
