% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_example.R
\name{sim_sleuth_data}
\alias{sim_sleuth_data}
\title{Generate an artificial minimal sleuth-like structure, for code-testing and for examples.}
\usage{
sim_sleuth_data(varname = "condition", COUNTS_COL = "est_counts",
  TARGET_COL = "target_id", PARENT_COL = "parent_id",
  BS_TARGET_COL = "target_id", cnames = c("one", "two"))
}
\arguments{
\item{varname}{Name for the variables by which to compare.}

\item{COUNTS_COL}{Name for the bootdtraps column containing counts.}

\item{TARGET_COL}{Name for annotation column containing transcript identification.}

\item{PARENT_COL}{Name for annotation column containing respective gene identification.}

\item{BS_TARGET_COL}{Name for bootstraps column containing transcript identification.}

\item{cnames}{A vector of (two) name values for the comparison variable.}
}
\value{
a list with: a minimal sleuth-like object, a corresponding annotation data.frame.
}
\description{
The default values here should match the default values expected by calculate_DTU().
}

