% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{structure_data}
\alias{structure_data}
\title{Structure input data.}
\usage{
structure_data(tx_filter, steps, threads, count_data_A, count_data_B,
  boot_data_A, boot_data_B)
}
\arguments{
\item{tx_filter}{Pre-processed annotation.}

\item{steps}{Infered input type.}

\item{threads}{Threads.}

\item{count_data_A}{Non-bootstrapped data for one condition.}

\item{count_data_B}{Non-bootstrapped data for other condition.}

\item{boot_data_A}{Boostrapped data for one condition.}

\item{boot_data_B}{Bootstrapped data for other condition.}
}
\value{
List.
}
\description{
Structure input data.
}
