% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\docType{class}
\name{TxtProgressUpdate-class}
\alias{TxtProgressUpdate}
\alias{TxtProgressUpdate-class}
\title{An S4 text-output progress update class, subclassed from ProgressUpdate}
\description{
An S4 text-output progress update class, subclassed from ProgressUpdate
}
\section{Slots}{

\describe{
\item{\code{step}}{The current position in the list of progress points}
}}

