% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\docType{methods}
\name{update,BarProgressUpdate-method}
\alias{update,BarProgressUpdate-method}
\title{Update progress by extending progress bar}
\usage{
\S4method{update}{BarProgressUpdate}(theObject)
}
\description{
Update progress by extending progress bar
}

