% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\docType{methods}
\name{update,TxtProgressUpdate-method}
\alias{update,TxtProgressUpdate-method}
\title{Update progress with next text string and percent complete figure}
\usage{
\S4method{update}{TxtProgressUpdate}(theObject)
}
\description{
Update progress with next text string and percent complete figure
}

