% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\docType{methods}
\name{update_progress,BarProgressUpdate-method}
\alias{update_progress,BarProgressUpdate-method}
\title{Update progress by extending progress bar}
\usage{
\S4method{update_progress}{BarProgressUpdate}(theObject)
}
\arguments{
\item{theObject}{the progress bar}
}
\description{
Update progress by extending progress bar
}

