% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\docType{methods}
\name{update_progress,TxtProgressUpdate-method}
\alias{update_progress,TxtProgressUpdate-method}
\title{Update progress with next text string and percent complete figure}
\usage{
\S4method{update_progress}{TxtProgressUpdate}(theObject)
}
\arguments{
\item{theObject}{The progress object.}
}
\description{
Update progress with next text string and percent complete figure
}

