% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{update_progress}
\alias{update_progress}
\title{Update progress reporting}
\usage{
update_progress(pb, progress_steps)
}
\arguments{
\item{pb}{A progress bar}

\item{progress_steps}{Vector of values to use as progress bar updates
(should match number of calls being made to update_progress)}
}
\description{
Update progress reporting
}

