% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{alloc_out}
\alias{alloc_out}
\title{Create output structure.}
\usage{
alloc_out(annot, full)
}
\arguments{
\item{annot}{Pre-processed by \code{tidy_annot()}.}

\item{full}{Full-sized structure or core fields only. Either "full" or "short".}
}
\value{
A list.
}
\description{
Create output structure.
}
