% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{annot2models}
\alias{annot2models}
\title{Prepare GTF/GFF3 for gene model plotting.}
\usage{
annot2models(annotfile, threads = detectCores())
}
\arguments{
\item{annotfile}{A GTF file.}

\item{threads}{Number of processing threads (all available).}
}
\value{
A list (by gene) of GRangeLists (by transcript).
}
\description{
Create GRanges for each transcript and bind them into a GRangeList for each gene.
Then one can easily `ggbio::autoplot()` a gene's models by its gene ID.
}
\details{
For GFF3, genes must have type 'gene' or 'SO:0000704' in order to be recognised.
For GTF, genes are infered as the unique values of gene_id.
}
