% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rats.R
\name{calculate_DTU}
\alias{calculate_DTU}
\title{Calculate differential transcript usage.}
\usage{
calculate_DTU(sleuth_data, transcripts, counts_col = "est_counts")
}
\arguments{
\item{sleuth_data}{A sleuth object}

\item{transcripts}{A dataframe listing the transcripts to process, and their parent genes,
with at least column \code{target_id} & \code{parent_id}}

\item{counts_col}{The sleuth column to use for the calculation (est_counts or tpm), default est_counts}
}
\value{
(TODO) dataframe with a row for each transcript, indicating if it is DTU and giving p-value etc.
}
\description{
Calculate differential transcript usage.
}

