% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rats.R
\name{calculate_DTU}
\alias{calculate_DTU}
\title{Calculate differential transcript usage.}
\usage{
calculate_DTU(sleuth_data, transcripts, name_A, name_B, varname = "condition",
  counts_col = "est_counts", correction = "BH", p_thresh = 0.05,
  TARGET_ID = "target_id", PARENT_ID = "parent_id",
  BS_TARGET_ID = "target_id", verbose = FALSE)
}
\arguments{
\item{sleuth_data}{A sleuth object}

\item{transcripts}{A dataframe matching the transcript IDs to their corresponding gene IDs.}

\item{name_A}{The sleuth name for one condition.}

\item{name_B}{The sleuth name for the other condition.}

\item{varname}{The sleuth name of the covariate to which the two conditions belong, default \code{"condition"}.}

\item{counts_col}{The sleuth column to use for the calculation (est_counts or tpm), default \code{"est_counts"}.}

\item{correction}{p-value correction to apply, as defined in \code{stats::p.adjust.methods}, default \code{"BH"}.}

\item{p_thresh}{p-value threshold, default 0.05.}

\item{TARGET_ID}{The name of the transcript id column in transcripts object, default \code{"target_id"}}

\item{PARENT_ID}{The name of the parent id column in transcripts object, default \code{"parent_id"}.}

\item{BS_TARGET_ID}{The name of the transcript id column in sleuth bootstrap tables, default \code{"target_id"}.}

\item{verbose}{Whether to progress updates, default \code{FALSE}.}
}
\value{
List of data frames, with gene-level and transcript-level information.
}
\description{
Calculate differential transcript usage.
}

