% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rats.R
\name{calculate_DTU}
\alias{calculate_DTU}
\title{Calculate differential transcript usage.}
\usage{
calculate_DTU(sleuth_data, transcripts, name_A, name_B, varname = "condition",
  p_thresh = 0.05, count_thresh = 5, testmode = "both",
  correction = "BH", verbose = FALSE, boots = TRUE,
  threads = parallel::detectCores(), COUNTS_COL = "est_counts",
  TARGET_COL = "target_id", PARENT_COL = "parent_id",
  BS_TARGET_COL = "target_id")
}
\arguments{
\item{sleuth_data}{A sleuth object.}

\item{transcripts}{A dataframe matching the transcript identifiers to their corresponding gene identifiers.}

\item{name_A}{The name for one condition, as it appears in the \code{sample_to_covariates} table within the sleuth object.}

\item{name_B}{The name for the other condition, as it appears in the \code{sample_to_covariates} table within the sleuth object.}

\item{varname}{The name of the covariate to which the two conditions belong, as it appears in the \code{sample_to_covariates} table within the sleuth object. Default \code{"condition"}.}

\item{p_thresh}{The p-value threshold, default 0.05.}

\item{count_thresh}{Transcripts with fewer reads will be ignored (default 5).}

\item{testmode}{One of "G-test", "proportion-test", "both" (default both).}

\item{correction}{The p-value correction to apply, as defined in \code{stats::p.adjust.methods}, default \code{"BH"}.}

\item{verbose}{Display progress updates, default \code{FALSE}.}

\item{boots}{Bootstrap the p-values. Default TRUE.}

\item{threads}{Enable parallel processing. Default uses parallel::detectCores(). Try setting to 1 if you are having issues.}

\item{COUNTS_COL}{The name of the counts column to use for the DTU calculation (est_counts or tpm), default \code{"est_counts"}.}

\item{TARGET_COL}{The name of the transcript identifier column in the transcripts object, default \code{"target_id"}}

\item{PARENT_COL}{The name of the parent identifier column in the transcripts object, default \code{"parent_id"}.}

\item{BS_TARGET_COL}{The name of the transcript identifier column in the sleuth bootstrap tables, default \code{"target_id"}.}
}
\value{
List of data tables, with gene-level and transcript-level information.
}
\description{
Calculate differential transcript usage.
}

