% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rats.R
\name{calculate_DTU}
\alias{calculate_DTU}
\title{Calculate differential transcript usage.}
\usage{
calculate_DTU(sleuth_data, transcripts, name_A, name_B, varname = "condition",
  counts_col = "est_counts", correction = "BH", p_thresh = 0.05,
  verbose = FALSE, threads = parallel::detectCores(),
  TARGET_ID = "target_id", PARENT_ID = "parent_id",
  BS_TARGET_ID = "target_id")
}
\arguments{
\item{sleuth_data}{A sleuth object.}

\item{transcripts}{A dataframe matching the transcript identifiers to their corresponding gene identifiers.}

\item{name_A}{The name for one condition, as it appears in the \code{sample_to_covariates} table within the sleuth object.}

\item{name_B}{The name for the other condition, as it appears in the \code{sample_to_covariates} table within the sleuth object.}

\item{varname}{The name of the covariate to which the two conditions belong, as it appears in the \code{sample_to_covariates} table within the sleuth object. Default \code{"condition"}.}

\item{counts_col}{The name of the counts column to use for the DTU calculation (est_counts or tpm), default \code{"est_counts"}.}

\item{correction}{The p-value correction to apply, as defined in \code{stats::p.adjust.methods}, default \code{"BH"}.}

\item{p_thresh}{The p-value threshold, default 0.05.}

\item{verbose}{Display progress updates, default \code{FALSE}.}

\item{threads}{Enable parallel processing. Defaults to the value returned by parallel::detectCores(). Try setting to 1 if you are having issues.}

\item{TARGET_ID}{The name of the transcript identifier column in the transcripts object, default \code{"target_id"}}

\item{PARENT_ID}{The name of the parent identifier column in the transcripts object, default \code{"parent_id"}.}

\item{BS_TARGET_ID}{The name of the transcript identifier column in the sleuth bootstrap tables, default \code{"target_id"}.}
}
\value{
List of data tables, with gene-level and transcript-level information.
}
\description{
Calculate differential transcript usage.
}

