% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rats.R
\name{calculate_stats2}
\alias{calculate_stats2}
\title{Calculate row-wise statistics in a dataframe.}
\usage{
calculate_stats2(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
A dataframe containing the calculated statistics, one in each column
}
\description{
Calculate row-wise statistics in a dataframe.
}

