% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rats.R
\name{calculate_tx_proportions}
\alias{calculate_tx_proportions}
\title{Calculate the proportion of counts which are assigned to each transcript in a gene}
\usage{
calculate_tx_proportions(count_data, transcripts)
}
\arguments{
\item{count_data}{dataframe containing the counts from all bootstraps for one condition}
}
\value{
Proportion of counts which are assigned to each transcript in a gene (but currently just mean and variance per transcript)
}
\description{
Calculate the proportion of counts which are assigned to each transcript in a gene
}

