% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rats.R
\name{call_DTU}
\alias{call_DTU}
\title{Calculate differential transcript usage.}
\usage{
call_DTU(annot = NULL, TARGET_COL = "target_id", PARENT_COL = "parent_id",
  slo = NULL, name_A = "Condition-A", name_B = "Condition-B",
  varname = "condition", COUNTS_COL = "est_counts",
  BS_TARGET_COL = "target_id", count_data_A = NULL, count_data_B = NULL,
  boot_data_A = NULL, boot_data_B = NULL, p_thresh = 0.05,
  count_thresh = 5, dprop_thresh = 0.1, conf_thresh = 0.95,
  correction = "BH", testmode = "both", boots = "both", bootnum = 100L,
  description = NA_character_, verbose = TRUE)
}
\arguments{
\item{annot}{A data.frame matching the transcript identifiers to their corresponding gene identifiers. Any additional columns are allowed but ignored.}

\item{TARGET_COL}{The name of the transcript identifier column in the \code{annot} object. (Default \code{"target_id"})}

\item{PARENT_COL}{The name of the parent identifier column in the \code{annot} object. (Default \code{"parent_id"})}

\item{slo}{A Sleuth object.}

\item{name_A}{The name for one condition, as it appears in the \code{sample_to_covariates} table within the Sleuth object.}

\item{name_B}{The name for the other condition, as it appears in the \code{sample_to_covariates} table within the sleuth object.}

\item{varname}{The name of the covariate to which the two conditions belong, as it appears in the \code{sample_to_covariates} table within the sleuth object. (Default \code{"condition"}).}

\item{COUNTS_COL}{The name of the counts column to use for the DTU calculation (est_counts or tpm). (Default \code{"est_counts"})}

\item{BS_TARGET_COL}{The name of the transcript identifier column in the sleuth bootstrap tables. (Default \code{"target_id"})}

\item{count_data_A}{A data.table of estimated counts for condition A. One column per sample/replicate, one row per transcript. The first column should contain the transcript identifiers.}

\item{count_data_B}{A data.table of estimated counts for condition B. One column per sample/replicate, one row per transcript. The first column should contain the transcript identifiers.}

\item{boot_data_A}{A list of data.tables, one per sample/replicate of condition A. One bootstrap iteration's estimates per column, one transcript per row. The first column should contain the transcript identifiers.}

\item{boot_data_B}{A list of data.tables, one per sample/replicate of condition B. One bootstrap iteration's estimates per column, one transcript per row. The first column should contain the transcript identifiers.}

\item{p_thresh}{The p-value threshold, default 0.05.}

\item{count_thresh}{Minimum count of fragments per sample, in at least one of the conditions, for transcripts to be eligible for testing. (Default 5)}

\item{dprop_thresh}{Minimum change in proportion (effect size) of a transcript for it to be eligible to be significant. (Default 0.1)}

\item{conf_thresh}{Confidence threshold. The fraction of bootstrap iterations calling DTU required to have confidence in the final call. (Default 0.95) Ignored if no bootstraps.}

\item{correction}{The p-value correction to apply, as defined in \code{stats::p.adjust.methods}. (Default \code{"BH"})}

\item{testmode}{One of "genes", "transc", "both". (Default "both")}

\item{boots}{Bootstrap the p-values of either test. One of "genes", "transc", "both" or "none". (Default "both")}

\item{bootnum}{Number of bootstraps. (Default 100)}

\item{description}{Free-text description of the run. You can use this to add metadata to the results object. The results' description field can also be filled in after the run.}

\item{verbose}{Display progress updates. (Default \code{TRUE})}
}
\value{
List of data tables, with gene-level and transcript-level information.
}
\description{
There are three options for input:
\itemize{
 \item{A sleuth object. This requires the following parameters: \code{slo}, \code{name_A}, \code{name_B} and optionally \code{varname}, \code{COUNT_COL}, \code{BS_TARGET_COL}.}
 \item{Bootstrapped count estimates. This requires the following parameters: \code{boot_data_A} and \code{boot_data_B}. \code{name_A} and \code{name_B} can optionally be used to name the conditions.}
 \item{Count estimates. This requires the following parameters: \code{count_data_A} and \code{count_data_B}. \code{name_A} and \code{name_B} can optionally be used to name the conditions.}
}
}

