% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rats.R
\name{call_DTU}
\alias{call_DTU}
\title{Calculate differential transcript usage.}
\usage{
call_DTU(
  annot = NULL,
  TARGET_COL = "target_id",
  PARENT_COL = "parent_id",
  count_data_A = NULL,
  count_data_B = NULL,
  boot_data_A = NULL,
  boot_data_B = NULL,
  name_A = "Condition-A",
  name_B = "Condition-B",
  varname = "condition",
  use_sums = FALSE,
  p_thresh = 0.05,
  abund_thresh = 5,
  dprop_thresh = 0.2,
  correction = "BH",
  scaling = 1,
  testmode = "both",
  lean = TRUE,
  qboot = TRUE,
  qbootnum = 0L,
  qrep_thresh = 0.95,
  rboot = TRUE,
  rrep_thresh = 0.85,
  description = NA_character_,
  verbose = TRUE,
  threads = 1L,
  seed = NA_integer_,
  reckless = FALSE,
  dbg = "0"
)
}
\arguments{
\item{annot}{A data.table matching transcript identifiers to gene identifiers. Any additional columns are allowed but ignored.}

\item{TARGET_COL}{The name of the column for the transcript identifiers in \code{annot}. (Default \code{"target_id"})}

\item{PARENT_COL}{The name of the column for the gene identifiers in \code{annot}. (Default \code{"parent_id"})}

\item{count_data_A}{A data.table of estimated counts for condition A. One column per sample/replicate, one row per transcript. The first column should contain the transcript identifiers.}

\item{count_data_B}{A data.table of estimated counts for condition B. One column per sample/replicate, one row per transcript. The first column should contain the transcript identifiers.}

\item{boot_data_A}{A list of data.tables, one per sample/replicate of condition A. One bootstrap iteration's estimates per column, one transcript per row. The first column should contain the transcript identifiers.}

\item{boot_data_B}{A list of data.tables, one per sample/replicate of condition B. One bootstrap iteration's estimates per column, one transcript per row. The first column should contain the transcript identifiers.}

\item{name_A}{The name for one condition. (Default "Condition-A")}

\item{name_B}{The name for the other condition. (Default "Condition-B")}

\item{varname}{The name of the covariate to which the two conditions belong. (Default \code{"condition"}).}

\item{use_sums}{Use each transcript's sum of abundances across the replicates, instead of the means. Increases sensitivity, but also increases risk of false positives. RATs used only the sums up to 0.6.5 (inclusive). (Default FALSE)}

\item{p_thresh}{The p-value threshold. (Default 0.05)}

\item{abund_thresh}{Noise threshold. Minimum mean (across replicates) abundance for transcripts (and genes) to be eligible for testing. (Default 5)}

\item{dprop_thresh}{Effect size threshold. Minimum change in proportion of a transcript for it to be considered meaningful. (Default 0.20)}

\item{correction}{The p-value correction to apply, as defined in  \code{\link[stats]{p.adjust.methods}}. (Default \code{"BH"})}

\item{scaling}{A scaling factor or vector of scaling factors, to be applied to the abundances *prior* to any thresholding and testing. Useful for scaling TPMs (transcripts per 1 million reads) to the actual library sizes of the samples. If a vector is supplied, the order should correspond to the samples in group A followed by the samples in group B. WARNING: Improper use of the scaling factor will artificially inflate/deflate the significances obtained.}

\item{testmode}{One of \itemize{\item{"genes"}, \item{"transc"}, \item{"both" (default)}}.}

\item{lean}{Reduce memory footprint by not tracking mean/median/max/min/stdev for Dprop and pval across bootstrap iterations. The respective columns will be absent from the output structure. (Default TRUE)}

\item{qboot}{Bootstrap the DTU robustness against bootstrapped quantifications data. (Default \code{TRUE}) Ignored if input is \code{count_data}.}

\item{qbootnum}{Number of iterations for \code{qboot}. (Default 0) If 0, RATs will try to infer a value from the data.}

\item{qrep_thresh}{Reproducibility threshold for quantification bootsrapping. (Default 0.95)}

\item{rboot}{Bootstrap the DTU robustness against the replicates. Does *ALL* 1 vs 1 combinations. (Default \code{TRUE})}

\item{rrep_thresh}{Reproducibility threshold for replicate bootsrapping. (Default 0.85) With few replicates per condition, the reproducibility takes heavily quantized values. For 3x3, there are 9 possible 1v1 comparisons, and a consistency of 8/9 = 0.88.}

\item{description}{Free-text description of the run. You can use this to add metadata to the results object.}

\item{verbose}{Display progress updates and warnings. (Default \code{TRUE})}

\item{threads}{Number of threads to use. (Default 1) Multi-threading will be ignored on non-POSIX systems.}

\item{seed}{A numeric integer used to initialise the random number engine. Use this only if reproducible bootstrap selections are required. (Default NA)}

\item{reckless}{RATs normally aborts if any inconsistency is detected among the transcript IDs found in the annotation and the quantifications. Enabling reckless mode will downgrade this error to a warning and allow RATs to continue the run. Not recommended unless you know why the inconsistency exists and how it will affect the results. (Default FALSE)}

\item{dbg}{Debugging mode only. Interrupt execution at the specified flag-point. Used to speed up code-tests by avoiding irrelevant downstream processing. (Default 0: do not interrupt)}
}
\value{
List of mixed types. Contains a list of runtime settings, a table of gene-level results, a table of transcript-level results, and a list of two tables with the transcript abundaces.
}
\description{
There are two modes for input:
\itemize{
 \item{Bootstrapped count estimates. This requires the following parameters: \code{boot_data_A} and \code{boot_data_B}.}
 \item{Count estimates. This requires the following parameters: \code{count_data_A} and \code{count_data_B}.}
}
}
