% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangedat-gen.R
\name{countrange_sim}
\alias{countrange_sim}
\title{Generate artificial read-count data for simulated 2-transcript genes, 
systematically covering a broad range of proportions and magnitudes.}
\usage{
countrange_sim(propfrom = 0, propto = 1, propby = 0.01,
  magnitudes = c(1, 5, 10, 30, 100, 500, 1000))
}
\arguments{
\item{propfrom, }{propto, propby: Range and step of proportions for 1st transcript. Sibling transcript complemented from these.}

\item{magnitudes:}{vector of read-count sizes to use.}
}
\value{
list containing: \code{data} a minimal sleuth-like object.
                         \code{anno} a dataframe matching \code{target_id} to \code{parent_id}.
                         \code{sim} a dataframe with the simulation parameters per transcript.
}
\description{
For each level of transcript proportions in condition A, all proportion levels are
generated for condition B as possible new states. Therefore the number of transcripts 
generated is 2 * N^2 * M, where 2 is the number of gene isoforms per paramater combination, 
N is the number of proportion levels and M the number of magnitude levels.
}

