% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{denest_sleuth_boots}
\alias{denest_sleuth_boots}
\title{Extract bootstrap counts into a less nested structure.}
\usage{
denest_sleuth_boots(slo, annot, samples, COUNTS_COL = "tpm",
  BS_TARGET_COL = "target_id", TARGET_COL = "target_id",
  PARENT_COL = "parent_id", threads = 1)
}
\arguments{
\item{slo}{A sleuth object.}

\item{annot}{A data.frame matching transcript identfier to gene identifiers.}

\item{samples}{A numeric vector of samples to extract counts for.}

\item{COUNTS_COL}{The name of the column with the counts. (Default "tpm")}

\item{BS_TARGET_COL}{The name of the column with the transcript IDs. (Default "target_id")}

\item{TARGET_COL}{The name of the column for the transcript identifiers in \code{annot}. (Default \code{"target_id"})}

\item{PARENT_COL}{The name of the column for the gene identifiers in \code{annot}. (Default \code{"parent_id"})}

\item{threads}{Number of threads. (Default 1)}
}
\value{
A list of data.tables, one per sample, containing all the bootstrap counts of the smaple. First column contains the transcript IDs.

NA replaced with 0.

Transcripts in \code{slo} that are missing from \code{annot} will be skipped completely.
Transcripts in \code{annot} that are missing from \code{slo} are automatically padded with NA, which we re-assign as 0.
}
\description{
*Legacy function*
}
\details{
It extracts the bootstrap data from the older-style \code{sleuth} object. 
As of sleuth version 0.29, the bootstrap data is no longer kept in the object.
}
