% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{dtu_plurality_summary}
\alias{dtu_plurality_summary}
\title{Summary of DTU plurality.}
\usage{
dtu_plurality_summary(dtuo)
}
\arguments{
\item{dtuo}{A DTU object.}
}
\value{
A named numerical vector giving a tally of the results.
}
\description{
Summary of DTU plurality.
}
