% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{dtu_summary}
\alias{dtu_summary}
\title{Summary of DTU calling.}
\usage{
dtu_summary(dtuo)
}
\arguments{
\item{dtuo}{A DTU object.}
}
\value{
A data.frame with category names in the first column and values in the second column.
}
\description{
A tally of DTU genes and transcripts.
}
