% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{dtu_switch_summary}
\alias{dtu_switch_summary}
\title{Summary of isoform switching events.}
\usage{
dtu_switch_summary(dtuo)
}
\arguments{
\item{dtuo}{A DTU object.}
}
\value{
A named numerical vector giving a tally of the results
}
\description{
Summary of isoform switching events.
}
