% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rats.R
\name{filter_and_match}
\alias{filter_and_match}
\title{Extract a filtered column from a bootstrap table}
\usage{
filter_and_match(bootstrap, tx_filter, counts_col, BS_TARGET_ID)
}
\arguments{
\item{bootstrap}{A bootstrap dataframe}

\item{tx_filter}{A boolean vector for filtering the bootstrap, with matching transcript ids}

\item{counts_col}{The column in the bootstrap table to extract}

\item{BS_TARGET_ID}{The name of transcript id column in sleuth bootstrap tables.}

\item{TARGET_ID}{The name of transcript id column in transcripts object.}
}
\value{
The column from the bootstrap table
}
\description{
Extract a filtered column from a bootstrap table
}

