% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_wrapper.R
\name{fish4rodents}
\alias{fish4rodents}
\title{Import abundances directly from salmon and kallisto output.}
\usage{
fish4rodents(A_paths, B_paths, annot, COUNTS_COL = "est_counts",
  BS_TARGET_COL = "target_id", TARGET_COL = "target_id",
  PARENT_COL = "parent_id", half_cooked = FALSE)
}
\arguments{
\item{A_paths}{(character) A vector of strings, listing the directory paths to the quantifications for the first condition. One directory per replicate. The directory name should be a unique identifier for the sample.}

\item{B_paths}{(character) A vector of strings, listing the directory paths to the quantifications for the second condition. One directory per replicate. The directory name should be a unique identifier for the sample.}

\item{annot}{(data.frame) A table matching transcript identifiers to gene identifiers. This should be the same that you will use with \code{call_DTU()}.}

\item{COUNTS_COL}{The name of the column with the counts. (Default "est_counts")}

\item{BS_TARGET_COL}{The name of the column with the transcript IDs. (Default "target_id")}

\item{TARGET_COL}{The name of the column for the transcript identifiers in \code{annot}. (Default \code{"target_id"})}

\item{PARENT_COL}{The name of the column for the gene identifiers in \code{annot}. (Default \code{"parent_id"})}

\item{half_cooked}{(logical) If TRUE, input is already in \code{Kallisto} h5 format and \code{wasabi} conversion will be skipped. Wasabi automatically skips conversion if an .h5 is present, so this parameter is redundant, unless wasabi is not installed. (Default FALSE)}
}
\value{
A list of two, representing the abundances per condition. These will be formatted in the RATs generic bootstrapped data input format.
}
\description{
Rather than re-inventing the wheel, this function is a wrapper for the wasabi converter,
the sleuth data loader and RATs' sleuth data extractor. This also has the advantage that the
sleuth loader scales TPMs to estimated transcript counts, which is the appropriate abundance
type for RATs. TPMs directly are not suitable, neither are read counts.
}
\details{
Converting, normalising and importing multiple bootstrapped abundance files takes a bit of time.

\code{wasabi} automatically skips format conversion if a folder already contains an \code{abundance.h5} file.
}
