% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{fish4rodents}
\alias{fish4rodents}
\title{Import abundances directly from salmon and kallisto output.}
\usage{
fish4rodents(A_paths, B_paths, annot, TARGET_COL = "target_id",
  PARENT_COL = "parent_id", half_cooked = FALSE, threads = 1L,
  scaleto = 1e+06)
}
\arguments{
\item{A_paths}{(character) A vector of strings, listing the directory paths to the quantifications for the first condition. One directory per replicate. The directory name should be a unique identifier for the sample.}

\item{B_paths}{(character) A vector of strings, listing the directory paths to the quantifications for the second condition. One directory per replicate. The directory name should be a unique identifier for the sample.}

\item{annot}{(data.frame) A table matching transcript identifiers to gene identifiers. This should be the same that you used for quantification and that you will use with \code{call_DTU()}. It is used to order the transcripts consistently throughout RATs.}

\item{TARGET_COL}{The name of the column for the transcript identifiers in \code{annot}. (Default \code{"target_id"})}

\item{PARENT_COL}{The name of the column for the gene identifiers in \code{annot}. (Default \code{"parent_id"})}

\item{half_cooked}{(logical) If TRUE, input is already in \code{Kallisto} h5 format and \code{wasabi} conversion will be skipped. Wasabi automatically skips conversion if abundance.h5 is present, so this parameter is redundant, unless wasabi is not installed. (Default FALSE)}

\item{threads}{(integer) For parallel processing. (Default 1)}

\item{scaleto}{(double) Scaling factor for normalised abundances. (Default 1000000 gives TPM)}
}
\value{
A list of two, representing the TPM abundances per condition. These will be formatted in the RATs generic bootstrapped data input format.
}
\description{
Uses \code{wasabi} to convert Salmon read counts format to Kallisto RHDF5 format,
then applies TPM normalisation using the info available from the abundance.h5 files.
}
\details{
Converting, normalising and importing multiple bootstrapped abundance files takes a bit of time.

\code{wasabi} automatically skips format conversion if a folder already contains an \code{abundance.h5} file.
}
