% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.test.R
\name{g.test}
\alias{g.test}
\title{Log-likelihood tests of independence & goodness of fit
Does Williams' and Yates' correction
does Monte Carlo simulation of p-values, via gtestsim.c}
\usage{
g.test(x, y = NULL, correct = "none", p = rep(1/length(x), length(x)),
  simulate.p.value = FALSE, B = 2000)
}
\arguments{
\item{x}{a numeric vector or matrix. x and y can also both be factors.}

\item{y}{a numeric vector; ignored if x is a matrix. If x is a factor, y should be a factor of the same length.}

\item{correct}{a logical indicating whether to apply continuity correction when computing the test statistic for 2 by 2 tables: one half is subtracted from all |O - E| differences; however, the correction will not be bigger than the differences themselves. No correction is done if simulate.p.value = TRUE.}

\item{p}{a vector of probabilities of the same length of x. An error is given if any entry of p is negative.}

\item{simulate.p.value}{a logical indicating whether to compute p-values by Monte Carlo simulation.}

\item{B}{an integer specifying the number of replicates used in the Monte Carlo test.'

G & q calculation from Sokal & Rohlf (1995) Biometry 3rd ed.
TOI Yates' correction taken from Mike Camann's 2x2 G-test fn.
GOF Yates' correction as described in Zar (2000)
more stuff taken from ctest's chisq.test()

Sourced and adapted from from:
V3.3 Pete Hurd Sept 29 2001. phurd@ualberta.ca
http://www.psych.ualberta.ca/~phurd/cruft/g.test.r}
}
\description{
Log-likelihood tests of independence & goodness of fit
Does Williams' and Yates' correction
does Monte Carlo simulation of p-values, via gtestsim.c
}

