% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{get_dtu_ids}
\alias{get_dtu_ids}
\title{List of DTU ids.}
\usage{
get_dtu_ids(dtuo)
}
\arguments{
\item{dtuo}{A DTU object.}
}
\value{
A list of vectors.
}
\description{
Get the IDs for DTU/nonDTU/NA genes and transcripts.
The IDs will be ordered by effect size.
}
