% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{group_samples}
\alias{group_samples}
\title{Group samples by covariate value.}
\usage{
group_samples(covariates)
}
\arguments{
\item{covariates}{A dataframe with different factor variables. Like the \code{sample_to_covariates} table of a \code{sleuth} object. Each row is a sample, each column is a covariate, each cell is a covariate value for the sample.}
}
\value{
list of lists (per covariate) of vectors (per factor level).
}
\description{
*Legacy function* No longer in use, but kept for possible general utility.
}
\details{
Converts a table where each row is a sample and each column a variable into a
lookup structure that matches each value of each variable to a vector of corresponding samples.
}
