% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{group_samples}
\alias{group_samples}
\title{Group sample numbers by factor.}
\usage{
group_samples(covariates)
}
\arguments{
\item{covariates}{A dataframe with different factor variables.}
}
\value{
list of lists (per covariate) of vectors (per factor level).

Row number corresponds to sample number.
}
\description{
Group sample numbers by factor.
}
