% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{gtf2ids}
\alias{gtf2ids}
\title{Extract matching transcript and gene IDs from a GTF file.}
\usage{
gtf2ids(annotfile, transc_header = "target_id", gene_header = "parent_id")
}
\arguments{
\item{annotfile}{A GTF file.}

\item{transc_header}{The title for the transcripts column in the output. (target_id)}

\item{gene_header}{The title for the genes column in the output. (parent_id)}
}
\value{
A data.table with two columns, matching transcript IDs to gene IDs.
}
\description{
Previously named annot2ids(). The old name will be discontinued.
}
\details{
GFF3 not supported.
}
