% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{init_progress}
\alias{init_progress}
\title{Initialise progress reporting}
\usage{
init_progress(progress_steps)
}
\arguments{
\item{progress_steps}{Vector of values to use as progress bar updates
(should match number of calls being made to update_progress)}
}
\description{
Initialise progress reporting
}

