% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\docType{methods}
\name{initialize,ProgressUpdate-method}
\alias{initialize,ProgressUpdate-method}
\title{Initialisation of progress update object}
\usage{
\S4method{initialize}{ProgressUpdate}(.Object, ..., steps, on = FALSE)
}
\arguments{
\item{.Object}{base class}

\item{...}{unnamed arguments}

\item{steps}{dataframe of steps to progress through}

\item{on}{whether the progress updater is switched on}
}
\description{
Initialisation of progress update object
}

