% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{insert_steps}
\alias{insert_steps}
\title{Insert steps generic}
\usage{
insert_steps(theObject, newsteps)
}
\arguments{
\item{theObject}{The progress object}

\item{newsteps}{New steps to insert to the progress steps}
}
\description{
Insert steps generic
}

