% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rats.R
\name{mark_sibling_targets2}
\alias{mark_sibling_targets2}
\title{Compute a logical vector marking as FALSE the single-target parents in a data frame.
(alternative implementation)}
\usage{
mark_sibling_targets2(ids, duptx = FALSE)
}
\arguments{
\item{ids}{a data frame with at least two variables, \code{target_id} & \code{parent_id}}

\item{duptx}{a boolean switch indicating whether to account for duplicate target_ids (default FALSE)}
}
\value{
data.frame An updated version of the input ids.

Target IDs are assumed to have the format \code{parentID.childextension}.
}
\description{
Compute a logical vector marking as FALSE the single-target parents in a data frame.
(alternative implementation)
}

