% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{maxabs}
\alias{maxabs}
\title{Get largest value by absolute comparison.}
\usage{
maxabs(v)
}
\arguments{
\item{v}{A numeric vector.}
}
\value{
A numeric value.
}
\description{
Get back the original signed value. In case of equal absolutes, the positive
value will be returned.
}
