% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{parameters_are_good}
\alias{parameters_are_good}
\title{Check input parameters.}
\usage{
parameters_are_good(slo, annot, name_A, name_B, varname, COUNTS_COL, correction,
  p_thresh, TARGET_COL, PARENT_COL, BS_TARGET_COL, count_thresh, testmode,
  boots, bootnum, dprop_thresh, count_data_A, count_data_B, boot_data_A,
  boot_data_B, conf_thresh, threads)
}
\arguments{
\item{slo}{sleuth object}

\item{annot}{annotation dataframe}

\item{name_A}{condition name}

\item{name_B}{condition name}

\item{varname}{name of condition variable}

\item{COUNTS_COL}{name of counts column in bootstrap}

\item{correction}{p-value correction method}

\item{p_thresh}{significance level}

\item{TARGET_COL}{name of transcript id column in annotation}

\item{PARENT_COL}{name of gene id column in annotation}

\item{BS_TARGET_COL}{name of transcript id column in bootstrap}

\item{count_thresh}{minimum frgments per transcript per sample}

\item{testmode}{which tests to run}

\item{boots}{which tests to bootstrap}

\item{bootnum}{number of bootstrap iterations}

\item{dprop_thresh}{minimum change in proportion}

\item{count_data_A}{A dataframe of estimated counts.}

\item{count_data_B}{A dataframe of estimated counts.}

\item{boot_data_A}{A list of dataframes, one per sample, each with all the bootstrapped estimetes for the sample.}

\item{boot_data_B}{A list of dataframes, one per sample, each with all the bootstrapped estimetes for the sample.}

\item{conf_thresh}{Confidence threshold.}

\item{threads}{Number of threads.}
}
\value{
List: \itemize{
 \item{"error"}{logical}
 \item{"message"}{string}
 \item{"maxboots"}{Infered rule-of-thumbs number of iterations to do.}
 \item{"warn"}{logical}
 \item{"warnings}{list of strings}
}
}
\description{
Check input parameters.
}

