% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{parameters_are_good}
\alias{parameters_are_good}
\title{Check input parameters.}
\usage{
parameters_are_good(annot, count_data_A, count_data_B, boot_data_A, boot_data_B,
  TARGET_COL, PARENT_COL, correction, testmode, scaling, threads, p_thresh,
  abund_thresh, dprop_thresh, qboot, qbootnum, qrep_thresh, rboot, rrep_thresh)
}
\arguments{
\item{annot}{Annotation dataframe.}

\item{count_data_A}{A dataframe of estimated counts.}

\item{count_data_B}{A dataframe of estimated counts.}

\item{boot_data_A}{A list of dataframes, one per sample, each with all the bootstrapped estimetes for the sample.}

\item{boot_data_B}{A list of dataframes, one per sample, each with all the bootstrapped estimetes for the sample.}

\item{TARGET_COL}{Name of transcript id column in annotation.}

\item{PARENT_COL}{Name of gene id column in annotation.}

\item{correction}{P-value correction method.}

\item{testmode}{Which test to run.}

\item{scaling}{Abundance scaling factor.}

\item{threads}{Number of threads.}

\item{p_thresh}{Significance level.}

\item{abund_thresh}{Minimum transcript abundance per sample.}

\item{dprop_thresh}{Minimum change in proportion.}

\item{qboot}{Whether to bootstrap against quantifications.}

\item{qbootnum}{Number of bootstrap iterations.}

\item{qrep_thresh}{Confidence threshold.}

\item{rboot}{Whether to bootstrap against samples.}

\item{rrep_thresh}{Confidence threshold.}
}
\value{
List: \itemize{
 \item{"error"}{logical}
 \item{"message"}{string}
 \item{"maxboots"}{Infered rule-of-thumbs number of iterations to do.}
 \item{"warn"}{logical}
 \item{"warnings}{list of strings}
}
}
\description{
Check input parameters.
}
