% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rats.R
\name{parameters_good}
\alias{parameters_good}
\title{Check input parameters.}
\usage{
parameters_good(slo, annot, name_A, name_B, varname, COUNTS_COL, correction,
  p_thresh, TARGET_COL, PARENT_COL, BS_TARGET_COL, count_thresh, testmode,
  boots, bootnum, dprop_thresh)
}
\arguments{
\item{slo}{sleuth object}

\item{annot}{annotation dataframe}

\item{name_A}{condition name}

\item{name_B}{condition name}

\item{varname}{name of condition variable}

\item{COUNTS_COL}{name of counts column in bootstrap}

\item{correction}{p-value correction method}

\item{p_thresh}{significance level}

\item{TARGET_COL}{name of transcript id column in annotation}

\item{PARENT_COL}{name of gene id column in annotation}

\item{BS_TARGET_COL}{name of transcript id column in bootstrap}

\item{count_thresh}{minimum frgments per transcript per sample}

\item{testmode}{which tests to run}

\item{boots}{which tests to bootstrap}

\item{bootnum}{number of bootstrap iterations}

\item{dprop_thresh}{minimum change in proportion}
}
\value{
List with a logical value and a message.
}
\description{
Check input parameters.
}

