% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rats.R
\name{parameters_good}
\alias{parameters_good}
\title{Check input parameters.}
\usage{
parameters_good(slo, annot, ref_name, comp_name, varname, COUNTS_COL,
  correction, p_thresh, TARGET_COL, PARENT_COL, BS_TARGET_COL, verbose, threads,
  count_thresh, testmode, boots)
}
\value{
List with a logical value and a message.
}
\description{
Check input parameters.
}

