% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot_diagnostics}
\alias{plot_diagnostics}
\title{Plot diagnostics.}
\usage{
plot_diagnostics(dtuo, type = "cormat")
}
\arguments{
\item{dtuo}{A DTU object.}
}
\value{
A ggplot2 object. Simply display it or you can also customize it.
}
\description{
Plot diagnostics.
}
