% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_gene}
\alias{plot_gene}
\title{Plot count or proportion changes for all transcripts of a specified gene.}
\usage{
plot_gene(dtuo, pid, plt_type = "counts", plt_style = "bar")
}
\arguments{
\item{dtuo}{A DTU object.}

\item{pid}{A \code{parent_id} to make the plot for.}

\item{plt_type}{Values to plot. Either "counts" or "proportion". Default "counts".}

\item{plt_style}{Style of plot. Either "bars" or "lines". Default "bars".}
}
\description{
Plot count or proportion changes for all transcripts of a specified gene.
}

