% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot_gene}
\alias{plot_gene}
\title{Plot count and proportion changes for all transcripts of a specified gene.}
\usage{
plot_gene(dtuo, pid)
}
\arguments{
\item{dtuo}{A DTU object.}

\item{pid}{A \code{parent_id} to make the plot for.}
}
\value{
a ggplot2 object. Simply display it or you can also customize it.
}
\description{
Plot count and proportion changes for all transcripts of a specified gene.
}

