% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_gene}
\alias{plot_gene}
\title{Plot count or proportion changes for all transcripts of a specified gene.}
\usage{
plot_gene(dtuo, pid, vals = "counts", style = "bars")
}
\arguments{
\item{dtuo}{A DTU object.}

\item{pid}{A \code{parent_id} to make the plot for.}

\item{vals}{Values to plot. Either "counts" or "proportions". Default "counts".}

\item{style}{Style of plot. Either "bars" or "lines". Default "bars".}
}
\value{
a ggplot2 object. Simply display it or you can also customize it.

The error bars represent either the error of proportion or 2 standard deviations
of the mean count, according to the type of plot requested.
}
\description{
Plot count or proportion changes for all transcripts of a specified gene.
}

