% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot_overview}
\alias{plot_overview}
\title{Plot DTU results overview.}
\usage{
plot_overview(dtuo, type = "volcano")
}
\arguments{
\item{dtuo}{A DTU object.}

\item{type}{Type of plot. \itemize{
\item{"tvolcano" - Change in proportion VS. transcript-level statistical significance. (Default)}
\item{"gvolcano" - Largest change in proportion per gene VS. gene-level statistical significance.}
\item{"maxdprop" - Distribution of biggest change in proportion in each gene.}
\item{"dprop" - Distribution of effect size.}
\item{"reprod" - Distribution of gene-level DTU reproducibility.}
\item{"reprodVSdprop" - Transcript-level quantification reproducibility VS effect size.}
\item{"fcvolcano" - Fold-change in abundance VS. statistical significance. Done at the transcript level.}
\item{"fcVSdprop" - Fold-change of abundance VS difference in proportion, ofr each transcript.}}}
}
\value{
A ggplot2 object. Simply display it or you can also customize it.

CAUTION: RATs does NOT normalise the input abundances for fold-change calclulations. RATs
is NOT intended for study of DTE.

These overviews rely on the results of the transcript-level proportion tests. If your DTU
object was created without the transcript-level tests, this function will not work.
}
\description{
Plot DTU results overview.
}
