% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_overview}
\alias{plot_overview}
\title{Plot DTU results from the proportions test.}
\usage{
plot_overview(dtuo, type = "dpropVsig")
}
\arguments{
\item{dtuo}{A DTU object.}

\item{type}{Type of plot: "propVcount", "dpropVcount", "dpropVsig", "maxdprop"}
}
\value{
a ggplot2 object. Simply display it or you can also customize it.

Generally uses the results of the transcript-level proportion tests.
}
\description{
Plot DTU results from the proportions test.
}

