% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot_overview}
\alias{plot_overview}
\title{Plot DTU results overview.}
\usage{
plot_overview(dtuo, type = "volcano")
}
\arguments{
\item{dtuo}{A DTU object.}

\item{type}{Type of plot. \itemize{
\item{"volcano" - Change in proportion VS. statistical significance. Done at the transcript level. (Default)}
\item{"tradvolcano" - Fold-change in abundance VS. statistical significance. Done at the transcript level.}
\item{"maxdprop" - Distribution of biggest change in proportion in each gene.}
\item{"maxfc" - Distribution of biggest fold-change in isoform abundance in each gene.} }}
}
\value{
A ggplot2 object. Simply display it or you can also customize it.

These overviews rely on the results of the transcript-level proportion tests. If your DTU
object was created without the transcript-level tests, this function will not work.
}
\description{
Plot DTU results overview.
}
