% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot_overview}
\alias{plot_overview}
\title{Plot DTU results from the proportions test.}
\usage{
plot_overview(dtuo, type = "volcano")
}
\arguments{
\item{dtuo}{A DTU object.}

\item{type}{Type of plot. \itemize{
  \item{"volcano"}{Change in proportion VS. statistical significance. Done at the transcript level. (Default)}
  \item{"maxdprop"}{Distribution of biggest change in proportion in each gene.}
  \item{"transc_conf"}{Transcript-level confidence threshold VS. number of DTU positive calls.}
  \item{"gene_conf"}{Gene-level confidence threshold VS. number of DTU positive calls.}
}}
}
\value{
a ggplot2 object. Simply display it or you can also customize it.

Generally uses the results of the transcript-level proportion tests.
}
\description{
Plot DTU results from the proportions test.
}

