% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangedat-gen.R
\name{plot_sim}
\alias{plot_sim}
\title{Plot relationship of parameters.}
\usage{
plot_sim(data, type = "AvBvM")
}
\arguments{
\item{data}{the output from combine_sim_dtu()}

\item{type}{(str) "AvBvM", "B/AvM", "AvBvMprop", "B/AvMprop", "B-AvM"}
}
\description{
Plot relationship of parameters.
}

