% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotDTU}
\alias{plotDTU}
\title{Plot DTU results.}
\usage{
plotDTU(dtuo, type = "propVmag")
}
\arguments{
\item{dtuo}{A DTU object.}

\item{type}{Type of plot: "propVmag", "dpropVmag", "volcano", "dprop"}
}
\description{
Plot DTU results.
}

