% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotGeneDTU}
\alias{plotGeneDTU}
\title{Plot count or proportion DTU results for a specified gene}
\usage{
plotGeneDTU(DTUdata, pid, nreps = 7, ptype = "counts")
}
\arguments{
\item{DTUdata}{a DTU object with the data in to visualize.}

\item{pid}{a \code{parent_id} to make the plot for.}

\item{nreps}{the number of biological replicates per condition (this is used to calculate the error-bars for the 'proportions' plot). The default is 7.}

\item{ptype}{a switch for plotting either \code{"counts"} or \code{"proportion"}. The default is "counts".}
}
\description{
Plot count or proportion DTU results for a specified gene
}

