% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangedat-gen.R
\name{rangedat_gen}
\alias{rangedat_gen}
\title{Generate artificial 2-transcript-genes read-count data, systematically covering a broad range of proportions and magnitudes.}
\usage{
rangedat_gen()
}
\value{
sleuth-like object. Contains \code{"kal"} ans \code{"bootstrap"} with the simulated data and the non-sleuth 
        \code{"rangesim"} with the simulation details.
}
\description{
Generate artificial 2-transcript-genes read-count data, systematically covering a broad range of proportions and magnitudes.
}

