% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulators.R
\name{sim_boot_data}
\alias{sim_boot_data}
\title{Generate an artificial dataset of bootstrapped abundance estimates, for code-testing or examples.}
\usage{
sim_boot_data(cnames = c("A", "B"))
}
\arguments{
\item{cnames}{A vector of (two) name values for the comparison variable.}
}
\value{
A list with 3 elements. First a data.frame with the corresponding annotation. Then, 2 lists of data.tables. One list per condition. Each data table represents a sample and contains the estimates from the bootstrap iterations.
}
\description{
Based on sim_sleuth_data().
}
