% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulators.R
\name{sim_count_data}
\alias{sim_count_data}
\title{Generate an artificial dataset of abundance estimates, for code-testing or examples.}
\usage{
sim_count_data(PARENT_COL = "parent_id", TARGET_COL = "target_id",
  clean = TRUE)
}
\arguments{
\item{PARENT_COL}{Name for the bootdtraps column containing counts.}

\item{TARGET_COL}{Name for annotation column containing transcript identification.}

\item{clean}{Logical. Remove the intentional inconsistency of IDs between annotation and abundances.}
}
\value{
A list with 3 elements. First, a data.frame with the corresponding annotation table. 
Then, 2 data.tables, one per condition. Each data table contains the abundance estimates for all the samples for the respective condition.
}
\description{
Generate an artificial dataset of abundance estimates, for code-testing or examples.
}
